refactor: split main.ts
