+++ apps/widget-configurator/src/theme/commonTypography.ts
@@ -1,0 +1,8 @@
+import { TypographyOptions } from '@mui/material/styles/createTypography'
+
+export const commonTypography: TypographyOptions = {
+  htmlFontSize: 10,
+  button: {
+    textTransform: 'none' as const,
+  },
+}
